# Sped up SSSR by not storing every path back to root