Before this commit:

3EOH: 0.72s
2J3N: 0.26s
1NKS: 0.018s

After this commit:
3EOH: 0.35s
2J3N: 0.07s
1NKS: 0.007s

<!--
Thanks for contributing a pull request! 
-->
#### Reference Issue
<!-- Example: Fixes #1234 -->


#### What does this implement/fix? Explain your changes.

This change speeds up ring performance by not storing every path back to the root. Instead, it keeps track of parents and rebuilds paths from the parents once a cycle is found. It also stops the BFS once the depth of the BFS is larger than the smallest ring (i.e., we found a path that is longer than the smallest ring).

#### Any other comments?