@greglandrum 2020-08-13
@rachelnwalker : thanks for the contribution!
The performance improvements are quite impressive and I'm looking forward to having this improvement in the RDKit core. Having this central piece of code go faster is going to make many people very happy.

There are a bunch of test failures in the CI builds that need to be resolved before the PR can be merged. I looked at a few and it seems that they are due to the fact that your new implementation returns the rings with the atoms/bonds ordered differently. Here's an example of that.

Using v2020_03_4:
```
n [2]: m =Chem.MolFromSmiles('CC1CCCC1')                                                              

In [3]: ri = m.GetRingInfo()                                                                           

In [5]: list(ri.AtomRings())                                                                           
Out[5]: [(1, 2, 3, 4, 5)]

In [6]: list(ri.BondRings())                                                                           
Out[6]: [(1, 2, 3, 4, 5)]
```
Using your branch:
```
In [2]: m =Chem.MolFromSmiles('CC1CCCC1')                                                                                       

In [3]: ri = m.GetRingInfo()                                                                                                    

In [4]: list(ri.AtomRings())                                                                                                    
Out[4]: [(2, 3, 4, 5, 1)]

In [5]: list(ri.BondRings())                                                                                                    
Out[5]: [(2, 3, 4, 5, 1)]
```
To be clear: the results from your code are not wrong... they're just ordered differently.

There are two ways to proceed:
1. modify your code so that the tests pass (i.e. the ordering of the atoms/bonds in the individual rings corresponds to what's in the current code)
2. modify the expected results in the tests to match what your code produces

I have a small preference for maintaining consistent results across versions (i.e. approach 1.), but if that's either a lot of work or slows things down a lot, it's not the end of the world to go with 2. When doing that it's important to confirm that the changed results are, in fact, just due to the different atom ordering.

I haven't checked to see if the ordering of the rings themselves in molecules which have multiple rings differs, but it wouldn't be terribly surprising if that were the case. That may be harder to modify?