@rachelnwalker 2020-08-14
@greglandrum Thanks for such a quick response! I believe the issue was that I associated the root of the BFS with the second half of a ring instead of the first half. I just fixed this in the most commit by switching the while loop conditions that rebuild the ring from the parent pointers